requirements.txt
networkx
python-louvain
node2vec
scikit-learn
matplotlib
streamlit
pandas
python-igraph
leidenalg
